requirements.txt
asttokens==2.4.1
colorama==0.4.6
comm==0.2.2
comtypes==1.4.6
debugpy==1.8.5
decorator==5.1.1
dnspython==2.6.1
executing==2.0.1
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
PyAudio==0.2.14
Pygments==2.18.0
pymongo==4.8.0
pypiwin32==223
python-dateutil==2.9.0.post0
pyttsx3==2.91
pywin32==306
pyzmq==26.2.0
six==1.16.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
wcwidth==0.2.13
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.12
pymongo==4.5.0
ibm-watson==6.0.0
python-dotenv==1.0.0
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.12
pymongo==4.5.0
ibm-watson==6.0.0
python-dotenv==1.0.0
faiss-cpu==1.7.2
streamlit==1.18.0
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.12
ibm-watson==6.0.0
pymongo==4.5.0
faiss-cpu==1.7.2
streamlit==1.18.0
python-dotenv==1.0.0
numpy==1.24.3
